Fixing controllers and add edit controller function

+++ WebApplication3/Models/Service.cs
@@ -9,7 +9,7 @@ namespace WebApplication3.Models
     {
         public int Id { get; set; }
         public int AmountService { get; set; }
-        public decimal PaymentService { get; set; }
+        public int PaymentService { get; set; }
         public DateTime TripDate { get; set; }
 
     }
